requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
autorootcwd==1.0.1
billiard==4.2.1
celery==5.5.2
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
connected-components-3d==3.23.0
contourpy==1.3.2
cramjam==2.10.0
cryptography==44.0.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.8
dvc==3.59.2
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
dvclive==3.48.2
edt==2.4.1
einops==0.8.1
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
fastparquet==2024.11.0
fastremap==1.16.1
filelock==3.13.1
flatten-dict==0.4.2
flufl-lock==8.2.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.6.1
funcy==2.0
gitdb==4.0.12
gitpython==3.1.44
grandalf==0.8
grpcio==1.71.0
gto==1.7.2
h5py==3.13.0
huggingface-hub==0.31.2
hydra-callbacks==0.6.1
hydra-core==1.3.2
idna==3.4
imageio==2.37.0
imops==0.9.0
importlib-resources==6.5.2
ipykernel==6.29.5
ipython==8.36.0
iterative-telemetry==0.0.10
jedi==0.19.2
jinja2==3.1.4
joblib==1.5.0
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
kombu==5.5.3
lazy-loader==0.4
lightning==2.4.0
lightning-utilities==0.14.3
lxml==5.4.0
markdown==3.8
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
medimm==0.0.7
monai==1.4.0
mpmath==1.3.0
multidict==6.4.3
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.3.2
nilearn==0.11.1
numpy==1.26.4
nvidia-ml-py==12.575.51
omegaconf==2.3.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pydot==4.0.0
pygit2==1.18.0
pygments==2.19.1
pygtrie==2.5.0
pynrrd==1.1.3
pynvml==12.0.0
pyparsing==3.2.3
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1.post0
pytz==2025.2
pywin32==310
pyyaml==6.0.2
pyzmq==26.4.0
requests==2.28.1
rich==14.0.0
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.5.1
scipy==1.15.3
scmrepo==3.3.11
semver==3.0.4
setuptools==80.3.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simpleitk==2.5.0
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tomlkit==0.13.2
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchmetrics==1.7.2
torchvision==0.16.1+cu121
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.3
typing-extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==1.26.13
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
werkzeug==3.1.3
yarl==1.20.0
zc-lockfile==3.0.post1
